# Make some optimizations to the CMake build system