@nikic 2025-10-05
Please separate the individual changes here into separate PRs.