@efriedma-quic 2025-10-06
https://github.com/google/benchmark is a third-party project, and we prefer to minimize differences with upstream. Please submit changes there first.